+++ packages/eslint-config-react/plugins/next.js
@@ -1,12 +1,16 @@
 module.exports = {
   plugins: [
-    'next',
+    '@next/eslint-plugin-next',
   ],
   rules: {
+    'next/missing-preload': ['error'],
+
     'next/no-css-tags': ['error'],
 
     'next/no-html-link-for-pages': ['error'],
 
     'next/no-sync-scripts': ['error'],
+
+    'next/no-unwanted-polyfillio': ['error'],
   },
 }