+++ packages/eslint-config/.eslintrc.js
@@ -13,4 +13,7 @@ module.exports = {
     'quote-props': ['error', 'consistent-as-needed'], // since the majority of rules are in quotes with little outliers;
     'max-lines': ['off'], // We're not interested in breaking up rule files
   },
+  parserOptions: {
+    requireConfigFile: false,
+  },
 }