+++ packages/eslint-config/plugins/fuelrats.js
@@ -1,0 +1,8 @@
+module.exports = {
+  plugins: [
+    '@fuelrats',
+  ],
+  rules: {
+    '@fuelrats/no-mixed-access-class-members': ['error'],
+  },
+}