+++ packages/eslint-config/purejs.js
@@ -1,23 +1,18 @@
-/***************************************************************************\
-  All the credit and love to AirBnB for an amazing config project structure
-  https://github.com/airbnb/javascript
-\***************************************************************************/
-
 module.exports = {
   env: {
     es6: true,
   },
   extends: [
-    './rules/bestpractices',
-    './rules/errors',
-    './rules/es6',
-    './rules/style',
-    './rules/variables',
-    './rules/plugin-import',
-    './rules/plugin-jsdoc',
+    './core/bestpractices',
+    './core/errors',
+    './core/es6',
+    './core/style',
+    './core/variables',
+    './core/plugin-import',
+    './core/plugin-jsdoc',
   ].map(require.resolve),
   parserOptions: {
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: 'module',
   },
   reportUnusedDisableDirectives: true,