+++ packages/eslint-plugin/babel.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  plugins: [
+    '@babel/proposal-class-properties',
+    '@babel/proposal-private-methods',
+    '@babel/proposal-private-property-in-object',
+  ]
+}