+++ packages/eslint-plugin/util/RuleTester.js
@@ -1,0 +1,11 @@
+const { RuleTester } = require("eslint");
+
+RuleTester.setDefaultConfig({
+  parser: require.resolve('@babel/eslint-parser'),
+  parserOptions: {
+    sourceType: "module",
+    ecmaVersion: 2020,
+  },
+});
+
+module.exports = RuleTester;